chore: split huge shared.tsx file in affiliate/pure
