--- apps/cowswap-frontend/src/modules/affiliate/pure/AffiliateTraderModal/styles/index.ts
@@ -1,3 +1,0 @@
-export * from './layout'
-export * from './form'
-export * from './status'
